Refactor test fixtures
So that all projects using Testcontainers use the same base test
class.
