+++ test-resources-kafka/src/test/groovy/io/micronaut/testresources/kafka/AbstractKafkaSpec.groovy
@@ -1,25 +1,22 @@
 package io.micronaut.testresources.kafka
 
-import com.github.dockerjava.api.DockerClient
-import com.github.dockerjava.api.model.Container
+
 import io.micronaut.configuration.kafka.annotation.KafkaClient
 import io.micronaut.configuration.kafka.annotation.Topic
 import io.micronaut.context.annotation.Prototype
-import io.micronaut.test.support.TestPropertyProvider
-import io.micronaut.testresources.core.Scope
-import io.micronaut.testresources.testcontainers.TestContainers
-import org.testcontainers.DockerClientFactory
+import io.micronaut.testresources.testcontainers.AbstractTestContainersSpec
 import reactor.core.publisher.Mono
-import spock.lang.Specification
 
-abstract class AbstractKafkaSpec extends Specification implements TestPropertyProvider {
+abstract class AbstractKafkaSpec  extends AbstractTestContainersSpec {
 
-    Map<String, String> getProperties() {
-        [(Scope.PROPERTY_KEY): 'kafka']
+    @Override
+    String getScopeName() {
+        'kafka'
     }
 
-    void cleanupSpec() {
-        TestContainers.closeScope("kafka")
+    @Override
+    String getImageName() {
+        'cp-kafka'
     }
 
     @Prototype
@@ -33,24 +30,4 @@ abstract class AbstractKafkaSpec extends Specification implements TestPropertyPr
         Mono<String> updateAnalytics(String book);
     }
 
-    protected DockerClient dockerClient() {
-        DockerClientFactory.instance().client()
-    }
-
-    protected List<Container> runningTestContainers() {
-        dockerClient().listContainersCmd()
-                .exec()
-                .findAll {
-                    it.labels['org.testcontainers'] == 'true'
-                }
-                .findAll {
-                    println it
-                    it.state == 'running'
-                }
-    }
-
-    protected List<Container> kafkaContainers() {
-        runningTestContainers()
-                .findAll { it.image.contains('cp-kafka') }
-    }
 }
