+++ test-resources-kafka/src/test/groovy/io/micronaut/testresources/kafka/CustomKafkaImageTest.groovy
@@ -27,7 +27,7 @@ class CustomKafkaImageTest extends AbstractKafkaSpec {
         def result = client.updateAnalytics("oh yeah!")
 
         then:
-        kafkaContainers().size() == 1
+        listContainers().size() == 1
         result.block() == "oh yeah!"
         with(TestContainers.listByScope("kafka").get(Scope.of("kafka"))) {
             size() == 1
