+++ test-resources-mongodb/src/test/groovy/io/micronaut/testresources/mongodb/AbstractMongoDBSpec.groovy
@@ -1,41 +1,13 @@
 package io.micronaut.testresources.mongodb
 
-import com.github.dockerjava.api.DockerClient
-import com.github.dockerjava.api.model.Container
-import io.micronaut.test.support.TestPropertyProvider
-import io.micronaut.testresources.core.Scope
-import io.micronaut.testresources.testcontainers.TestContainers
-import org.testcontainers.DockerClientFactory
-import spock.lang.Specification
 
-abstract class AbstractMongoDBSpec extends Specification implements TestPropertyProvider {
+import io.micronaut.testresources.testcontainers.AbstractTestContainersSpec
 
-    Map<String, String> getProperties() {
-        [(Scope.PROPERTY_KEY): 'mongodb']
+abstract class AbstractMongoDBSpec extends AbstractTestContainersSpec {
+
+    @Override
+    String getScopeName() {
+        'mongodb'
     }
 
-    void cleanupSpec() {
-        TestContainers.closeScope("mongodb")
-    }
-
-    protected DockerClient dockerClient() {
-        DockerClientFactory.instance().client()
-    }
-
-    protected List<Container> runningTestContainers() {
-        dockerClient().listContainersCmd()
-                .exec()
-                .findAll {
-                    it.labels['org.testcontainers'] == 'true'
-                }
-                .findAll {
-                    println it
-                    it.state == 'running'
-                }
-    }
-
-    protected List<Container> mongodbContainers() {
-        runningTestContainers()
-                .findAll { it.image.contains('mongodb') }
-    }
 }
