+++ test-resources-neo4j/src/test/groovy/io/micronaut/testresources/neo4j/AbstractNeo4jDBSpec.groovy
@@ -1,45 +1,17 @@
 package io.micronaut.testresources.neo4j
 
-import com.github.dockerjava.api.DockerClient
-import com.github.dockerjava.api.model.Container
-import io.micronaut.test.support.TestPropertyProvider
-import io.micronaut.testresources.core.Scope
-import io.micronaut.testresources.testcontainers.TestContainers
-import jakarta.inject.Inject
-import org.testcontainers.DockerClientFactory
-import spock.lang.Specification
 
-abstract class AbstractNeo4jDBSpec extends Specification implements TestPropertyProvider {
+import io.micronaut.testresources.testcontainers.AbstractTestContainersSpec
+import jakarta.inject.Inject
+
+abstract class AbstractNeo4jDBSpec extends AbstractTestContainersSpec {
 
     @Inject
-    protected BookRepository bookRepository;
+    protected BookRepository bookRepository
 
-    Map<String, String> getProperties() {
-        [(Scope.PROPERTY_KEY): 'neo4j']
+    @Override
+    String getScopeName() {
+        'neo4j'
     }
 
-    void cleanupSpec() {
-        TestContainers.closeScope("neo4j")
-    }
-
-    protected DockerClient dockerClient() {
-        DockerClientFactory.instance().client()
-    }
-
-    protected List<Container> runningTestContainers() {
-        dockerClient().listContainersCmd()
-                .exec()
-                .findAll {
-                    it.labels['org.testcontainers'] == 'true'
-                }
-                .findAll {
-                    println it
-                    it.state == 'running'
-                }
-    }
-
-    protected List<Container> neo4jContainers() {
-        runningTestContainers()
-                .findAll { it.image.contains('neo4j') }
-    }
 }
